Clean up DC class
