+++ towhee/functional/data_collection.py
@@ -11,15 +11,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from typing import Any, Iterable, Iterator, Callable
-import random
+from typing import Iterable, Iterator, Callable
 import reprlib
 
 from towhee.functional.mixins.dag import register_dag
 from towhee.hparam import param_scope, dynamic_dispatch
-from towhee.functional.entity import Entity, EntityView
-from towhee.functional.option import Option, Some, Empty
+from towhee.functional.option import Option, Some
+from towhee.functional.entity import EntityView
 from towhee.functional.mixins import DCMixins
 from towhee.functional.mixins.dataframe import DataFrameMixin
 from towhee.functional.mixins.column import ColumnMixin
@@ -104,529 +102,6 @@ class DataCollection(Iterable, DCMixins):
         if hasattr(self._iterable, 'iterrows'):
             return (x[1] for x in self._iterable.iterrows())
         return iter(self._iterable)
-
-    @register_dag
-    def stream(self):
-        """
-        Create a stream data collection.
-
-        Examples:
-        1. Convert a data collection to streamed version
-
-        >>> dc = DataCollection([0, 1, 2, 3, 4])
-        >>> dc.is_stream
-        False
-
-        >>> dc = dc.stream()
-        >>> dc.is_stream
-        True
-        """
-        # pylint: disable=protected-access
-        iterable = iter(self._iterable) if not self.is_stream else self._iterable
-        return self._factory(iterable, parent_stream=False)
-
-    @register_dag
-    def unstream(self):
-        """
-        Create a unstream data collection.
-
-        Examples:
-
-        1. Create a unstream data collection
-
-        >>> dc = DataCollection(iter(range(5))).unstream()
-        >>> dc.is_stream
-        False
-
-        2. Convert a streamed data collection to unstream version
-
-        >>> dc = DataCollection(iter(range(5)))
-        >>> dc.is_stream
-        True
-        >>> dc = dc.unstream()
-        >>> dc.is_stream
-        False
-        """
-        iterable = list(self._iterable) if self.is_stream else self._iterable
-        return self._factory(iterable, parent_stream=False)
-
-    @property
-    def is_stream(self):
-        """
-        Check whether the data collection is stream or unstream.
-
-        Examples:
-
-        >>> dc = DataCollection([0,1,2,3,4])
-        >>> dc.is_stream
-        False
-
-        >>> result = dc.map(lambda x: x+1)
-        >>> result.is_stream
-        False
-        >>> result._iterable
-        [1, 2, 3, 4, 5]
-
-        >>> dc = DataCollection(iter(range(5)))
-        >>> dc.is_stream
-        True
-
-        >>> result = dc.map(lambda x: x+1)
-        >>> result.is_stream
-        True
-        >>> isinstance(result._iterable, Iterable)
-        True
-
-        """
-        return isinstance(self._iterable, Iterator)
-
-    def _factory(self, iterable, parent_stream=True):
-        """
-        Factory method for data collection.
-
-        This factory method has been wrapped into a `param_scope()` which contains parent information.
-
-        Args:
-            iterable: An iterable object, the data being stored in the DC
-            parent_stream: Whether to copy the parents format (streamed vs unstreamed)
-
-        Returns:
-            DataCollection: DataCollection encapsulating the iterable.
-        """
-        if parent_stream is True:
-            if self.is_stream:
-                if not isinstance(iterable, Iterator):
-                    iterable = iter(iterable)
-            else:
-                if isinstance(iterable, Iterator):
-                    iterable = list(iterable)
-
-        with param_scope() as hp:
-            hp().data_collection.parent = self
-            return DataCollection(iterable)
-
-    @register_dag
-    def exception_safe(self):
-        """
-        Making the data collection exception-safe by warp elements with `Option`.
-
-        Examples:
-
-        1. Exception breaks pipeline execution:
-
-        >>> dc = DataCollection.range(5)
-        >>> dc.map(lambda x: x / (0 if x == 3 else 2)).to_list()
-        Traceback (most recent call last):
-        ZeroDivisionError: division by zero
-
-        2. Exception-safe execution
-
-        >>> dc.exception_safe().map(lambda x: x / (0 if x == 3 else 2)).to_list()
-        [Some(0.0), Some(0.5), Some(1.0), Empty(), Some(2.0)]
-
-        >>> dc.exception_safe().map(lambda x: x / (0 if x == 3 else 2)).filter(lambda x: x < 1.5).to_list()
-        [Some(0.0), Some(0.5), Some(1.0), Empty()]
-
-        >>> dc.exception_safe().map(lambda x: x / (0 if x == 3 else 2)).filter(lambda x: x < 1.5, drop_empty=True).to_list()
-        [Some(0.0), Some(0.5), Some(1.0)]
-        """
-        result = map(lambda x: Some(x) if not isinstance(x, Option) else x, self._iterable)
-        return self._factory(result)
-
-    def safe(self):
-        """
-        Shortcut for `exception_safe`
-        """
-        return self.exception_safe()
-
-    @register_dag
-    def select_from(self, other):
-        """
-        Select data from dc with list(self).
-
-        Examples:
-
-        >>> dc1 = DataCollection([0.8, 0.9, 8.1, 9.2])
-        >>> dc2 = DataCollection([[1, 2, 0], [2, 3, 0]])
-
-        >>> dc3 = dc2.select_from(dc1)
-        >>> list(dc3)
-        [[0.9, 8.1, 0.8], [8.1, 9.2, 0.8]]
-        """
-        self.parent_ids.append(other.id)
-        other.notify_consumed(self.id)
-
-        def inner(x):
-            if isinstance(x, Iterable):
-                return [other[i] for i in x]
-            return other[x]
-
-        result = map(inner, self._iterable)
-        return self._factory(result)
-
-    @register_dag
-    def fill_empty(self, default: Any = None) -> 'DataCollection':
-        """
-        Unbox `Option` values and fill `Empty` with default values.
-
-        Args:
-            default (Any): default value to replace empty values;
-
-        Returns:
-            DataCollection: data collection with empty values filled with `default`;
-
-        Examples:
-
-        >>> dc = DataCollection.range(5)
-        >>> dc.safe().map(lambda x: x / (0 if x == 3 else 2)).fill_empty(-1.0).to_list()
-        [0.0, 0.5, 1.0, -1.0, 2.0]
-        """
-        result = map(lambda x: x.get() if isinstance(x, Some) else default, self._iterable)
-        return self._factory(result)
-
-    @register_dag
-    def drop_empty(self, callback: Callable = None) -> 'DataCollection':
-        """
-        Unbox `Option` values and drop `Empty`.
-
-        Args:
-            callback (Callable): handler for empty values;
-
-        Returns:
-            DataCollection: data collection that drops empty values;
-
-        Examples:
-
-        >>> dc = DataCollection.range(5)
-        >>> dc.safe().map(lambda x: x / (0 if x == 3 else 2)).drop_empty().to_list()
-        [0.0, 0.5, 1.0, 2.0]
-
-        Get inputs that case exceptions:
-
-        >>> exception_inputs = []
-        >>> result = dc.safe().map(lambda x: x / (0 if x == 3 else 2)).drop_empty(lambda x: exception_inputs.append(x.get().value))
-        >>> exception_inputs
-        [3]
-        """
-        if callback is not None:
-
-            def inner(data):
-                for x in data:
-                    if isinstance(x, Empty):
-                        callback(x)
-                    if isinstance(x, Some):
-                        yield x.get()
-
-            result = inner(self._iterable)
-        else:
-
-            def inner(data):
-                for x in data:
-                    if isinstance(x, Some):
-                        yield x.get()
-
-            result = inner(self._iterable)
-        return self._factory(result)
-
-    @register_dag
-    def map(self, *arg):
-        """
-        Apply operator to data collection.
-
-        Args:
-            *arg (Callable): functions/operators to apply to data collection;
-
-        Returns:
-            DataCollection: data collections that contains computation results;
-
-        Examples:
-
-        >>> dc = DataCollection([1,2,3,4])
-        >>> dc.map(lambda x: x+1).map(lambda x: x*2).to_list()
-        [4, 6, 8, 10]
-        """
-
-        # mmap
-        if len(arg) > 1:
-            return self.mmap(list(arg))
-        unary_op = arg[0]
-
-        # smap map for stateful operator
-        if hasattr(unary_op, 'is_stateful') and unary_op.is_stateful:
-            return self.smap(unary_op)
-
-        # pmap
-        if self.get_executor() is not None:
-            return self.pmap(unary_op)
-
-        if hasattr(self._iterable, 'map'):
-            return self._factory(self._iterable.map(unary_op))
-
-        if hasattr(self._iterable, 'apply') and hasattr(unary_op, '__dataframe_apply__'):
-            return self._factory(unary_op.__dataframe_apply__(self._iterable))
-
-        # map
-        def inner(x):
-            if isinstance(x, Option):
-                return x.map(unary_op)
-            else:
-                return unary_op(x)
-
-        result = map(inner, self._iterable)
-        return self._factory(result)
-
-    @register_dag
-    def zip(self, *others) -> 'DataCollection':
-        """
-        Combine two data collections.
-
-        Args:
-            *others (DataCollection): other data collections;
-
-        Returns:
-            DataCollection: data collection with zipped values;
-
-        Examples:
-
-        >>> dc1 = DataCollection([1,2,3,4])
-        >>> dc2 = DataCollection([1,2,3,4]).map(lambda x: x+1)
-        >>> dc3 = dc1.zip(dc2)
-        >>> list(dc3)
-        [(1, 2), (2, 3), (3, 4), (4, 5)]
-        """
-        self.parent_ids.extend([other.id for other in others])
-
-        for x in others:
-            x.notify_consumed(self.id)
-
-        return self._factory(zip(self, *others))
-
-    @register_dag
-    def filter(self, unary_op: Callable, drop_empty=False) -> 'DataCollection':
-        """
-        Filter data collection with `unary_op`.
-
-        Args:
-            unary_op (`Callable`):
-                Callable to decide whether to filter the element;
-            drop_empty (`bool`):
-                Drop empty values. Defaults to False.
-
-        Returns:
-            DataCollection: filtered data collection
-        """
-
-        # return filter(unary_op, self)
-        def inner(x):
-            if isinstance(x, Option):
-                if isinstance(x, Some):
-                    return unary_op(x.get())
-                return not drop_empty
-            return unary_op(x)
-
-        if hasattr(self._iterable, 'filter'):
-            return self._factory(self._iterable.filter(unary_op))
-
-        if hasattr(self._iterable, 'apply') and hasattr(unary_op, '__dataframe_filter__'):
-            return DataCollection(unary_op.__dataframe_apply__(self._iterable))
-
-        return self._factory(filter(inner, self._iterable))
-
-    @register_dag
-    def sample(self, ratio=1.0) -> 'DataCollection':
-        """
-        Sample the data collection.
-
-        Args:
-            ratio (float): sample ratio;
-
-        Returns:
-            DataCollection: sampled data collection;
-
-        Examples:
-
-        >>> dc = DataCollection(range(10000))
-        >>> result = dc.sample(0.1)
-        >>> ratio = len(result.to_list()) / 10000.
-        >>> 0.09 < ratio < 0.11
-        True
-        """
-        return self._factory(filter(lambda _: random.random() < ratio, self))
-
-    @staticmethod
-    @register_dag
-    def range(*arg, **kws):
-        """
-        Generate data collection with ranged numbers.
-
-        Examples:
-
-        >>> DataCollection.range(5).to_list()
-        [0, 1, 2, 3, 4]
-        """
-        return DataCollection(range(*arg, **kws))
-
-    @register_dag
-    def batch(self, size, drop_tail=False, raw=True):
-        """
-        Create small batches from data collections.
-
-        Args:
-            size (int): window size;
-            drop_tail (bool): drop tailing windows that not full, defaults to False;
-            raw (bool): whether to return raw data instead of DataCollection, defaults to True
-
-        Returns:
-            DataCollection of batched windows or batch raw data
-
-        Examples:
-
-        >>> dc = DataCollection(range(10))
-        >>> [list(batch) for batch in dc.batch(2, raw=False)]
-        [[0, 1], [2, 3], [4, 5], [6, 7], [8, 9]]
-
-        >>> dc = DataCollection(range(10))
-        >>> dc.batch(3)
-        [[0, 1, 2], [3, 4, 5], [6, 7, 8], [9]]
-
-        >>> dc = DataCollection(range(10))
-        >>> dc.batch(3, drop_tail=True)
-        [[0, 1, 2], [3, 4, 5], [6, 7, 8]]
-
-        >>> from towhee import Entity
-        >>> dc = DataCollection([Entity(a=a, b=b) for a,b in zip(['abc', 'vdfvcd', 'cdsc'], [1,2,3])])
-        >>> dc.batch(2)
-        [<Entity dict_keys(['a', 'b'])>, <Entity dict_keys(['a', 'b'])>]
-        """
-
-        def inner():
-            buff = []
-            count = 0
-            for ele in self._iterable:
-                if isinstance(ele, Entity):
-                    if count == 0:
-                        buff = ele
-                        for key in ele.__dict__.keys():
-                            buff.__dict__[key] = [buff.__dict__[key]]
-                        count = 1
-                        continue
-                    for key in ele.__dict__.keys():
-                        buff.__dict__[key].append(ele.__dict__[key])
-                else:
-                    buff.append(ele)
-                count += 1
-
-                if count == size:
-                    if raw:
-                        yield buff
-                    else:
-                        yield buff if isinstance(buff, list) else [buff]
-                    buff = []
-                    count = 0
-            if not drop_tail and count > 0:
-                if raw:
-                    yield buff
-                else:
-                    yield buff if isinstance(buff, list) else [buff]
-
-        return self._factory(inner())
-
-    @register_dag
-    def rolling(self, size: int, drop_head=True, drop_tail=True):
-        """
-        Create rolling windows from data collections.
-
-        Args:
-            size (int): window size;
-            drop_head (bool): drop headding windows that not full;
-            drop_tail (bool): drop tailing windows that not full;
-
-        Returns:
-            DataCollection: data collection of rolling windows;
-
-        Examples:
-
-        >>> dc = DataCollection(range(5))
-        >>> [list(batch) for batch in dc.rolling(3)]
-        [[0, 1, 2], [1, 2, 3], [2, 3, 4]]
-
-        >>> dc = DataCollection(range(5))
-        >>> [list(batch) for batch in dc.rolling(3, drop_head=False)]
-        [[0], [0, 1], [0, 1, 2], [1, 2, 3], [2, 3, 4]]
-
-        >>> dc = DataCollection(range(5))
-        >>> [list(batch) for batch in dc.rolling(3, drop_tail=False)]
-        [[0, 1, 2], [1, 2, 3], [2, 3, 4], [3, 4], [4]]
-        """
-
-        def inner():
-            buff = []
-            for ele in self._iterable:
-                buff.append(ele)
-                if not drop_head or len(buff) == size:
-                    yield buff.copy()
-                if len(buff) == size:
-                    buff = buff[1:]
-            while not drop_tail and len(buff) > 0:
-                yield buff
-                buff = buff[1:]
-
-        return self._factory(inner())
-
-    @register_dag
-    def flatten(self) -> 'DataCollection':
-        """
-        Flatten nested data collections.
-
-        Returns:
-            DataCollection: flattened data collection;
-
-        Examples:
-
-        >>> dc = DataCollection(range(10))
-        >>> nested_dc = dc.batch(2)
-        >>> nested_dc.flatten().to_list()
-        [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
-        """
-
-        def inner():
-            for ele in self._iterable:
-                if isinstance(ele, Iterable):
-                    for nested_ele in iter(ele):
-                        yield nested_ele
-                else:
-                    yield ele
-
-        return self._factory(inner())
-
-    @register_dag
-    def shuffle(self) -> 'DataCollection':
-        """
-        Shuffle an unstreamed data collection in place.
-
-        Returns:
-            DataCollection: shuffled data collection;
-
-        Examples:
-
-        1. Shuffle:
-
-        >>> dc = DataCollection([0, 1, 2, 3, 4])
-        >>> a = dc.shuffle()
-        >>> tuple(a) == tuple(range(5))
-        False
-
-        2. streamed data collection is not supported:
-
-        >>> dc = DataCollection([0, 1, 2, 3, 4]).stream()
-        >>> _ = dc.shuffle()
-        Traceback (most recent call last):
-        TypeError: shuffle is not supported for streamed data collection.
-        """
-        if self.is_stream:
-            raise TypeError('shuffle is not supported for streamed data collection.')
-        iterable = random.sample(self._iterable, len(self._iterable))
-        return self._factory(iterable)
 
     def __getattr__(self, name):
         """
@@ -728,180 +203,6 @@ class DataCollection(Iterable, DCMixins):
                 'data collection created from list or pandas DataFrame.')
         self._iterable[index] = value
 
-    def append(self, *args) -> 'DataCollection':
-        """
-        Append item to data collection.
-
-        Args:
-            item (Any): the item to append
-
-        Returns:
-            DataCollection: self
-
-        Examples:
-
-        >>> dc = DataCollection([0, 1, 2])
-        >>> dc.append(3).append(4)
-        [0, 1, 2, 3, 4]
-        """
-        if hasattr(self._iterable, 'append'):
-            self._iterable.append(*args)
-            return self
-        raise TypeError('append() is only supported for data collection created from list.')
-
-    def clear(self, *args) -> 'DataCollection':
-        """
-        Clear a DataCollection.
-
-        Examples:
-
-        >>> from towhee import DataCollection
-        >>> dc = DataCollection([1, 2, 3])
-        >>> dc.clear()
-        []
-        """
-        if hasattr(self._iterable, 'clear'):
-            self._iterable.clear(*args)
-            return self
-        raise TypeError('clear() is only supported for data collection created from list.')
-
-    def copy(self, *args) -> 'DataCollection':
-        """
-        Copy a DataCollection.
-
-        Examples:
-
-        >>> from towhee import DataCollection
-        >>> dc = DataCollection([1, 2, 3])
-        >>> dc_1 = dc.copy()
-        >>> dc_1._iterable.append(4)
-        >>> dc, dc_1
-        ([1, 2, 3], [1, 2, 3, 4])
-        """
-        if hasattr(self._iterable, 'copy'):
-            return self._factory(self._iterable.copy(*args))
-        raise TypeError('copy() is only supported for data collection created from list.')
-
-    def count(self, *args) -> int:
-        """
-        Count an element in  DataCollection.
-
-        Examples:
-
-        >>> from towhee import DataCollection
-        >>> dc = DataCollection([1, 2, 3])
-        >>> dc.count(1)
-        1
-        """
-        if hasattr(self._iterable, 'count'):
-            return self._iterable.count(*args)
-        raise TypeError('count() is only supported for data collection created from list.')
-
-    def extend(self, *args) -> 'DataCollection':
-        """
-        Extend a DataCollection.
-
-        Examples:
-
-        >>> from towhee import DataCollection
-        >>> dc = DataCollection([1, 2, 3])
-        >>> dc.extend([4, 5])
-        [1, 2, 3, 4, 5]
-        """
-        if hasattr(self._iterable, 'extend'):
-            self._iterable.extend(*args)
-            return self
-        raise TypeError('extend() is only supported for data collection created from list.')
-
-    def insert(self, *args) -> 'DataCollection':
-        """
-        Insert data into a DataCollection.
-
-        Examples:
-
-        >>> from towhee import DataCollection
-        >>> dc = DataCollection([1, 2, 3])
-        >>> dc.insert(0, 0)
-        [0, 1, 2, 3]
-        """
-        if hasattr(self._iterable, 'insert'):
-            self._iterable.insert(*args)
-            return self
-        raise TypeError('insert() is only supported for data collection created from list.')
-
-    def pop(self, *args) -> 'DataCollection':
-        """
-        Extend a DataCollection.
-
-        Examples:
-
-        >>> from towhee import DataCollection
-        >>> dc = DataCollection([1, 2, 3])
-        >>> dc.pop()
-        [1, 2]
-        """
-        if hasattr(self._iterable, 'pop'):
-            self._iterable.pop(*args)
-            return self
-        raise TypeError('pop() is only supported for data collection created from list.')
-
-    def remove(self, *args) -> 'DataCollection':
-        """
-        Remove element from DataCollection.
-
-        Examples:
-
-        >>> from towhee import DataCollection
-        >>> dc = DataCollection([1, 2, 3])
-        >>> dc.remove(1)
-        [2, 3]
-        """
-        if hasattr(self._iterable, 'remove'):
-            self._iterable.remove(*args)
-            return self
-        raise TypeError('remove() is only supported for data collection created from list.')
-
-    def reverse(self, *args) -> 'DataCollection':
-        """
-        Reverse a DataCollection.
-
-        Examples:
-
-        >>> from towhee import DataCollection
-        >>> dc = DataCollection([1, 2, 3])
-        >>> dc.reverse()
-        [3, 2, 1]
-        """
-        if hasattr(self._iterable, 'reverse'):
-            self._iterable.reverse(*args)
-            return self
-        raise TypeError('reverse() is only supported for data collection created from list.')
-
-    def sort(self, *args) -> 'DataCollection':
-        """
-        Sort a DataCollection.
-
-        Examples:
-
-        >>> from towhee import DataCollection
-        >>> dc = DataCollection([1, 4, 3])
-        >>> dc.sort()
-        [1, 3, 4]
-        """
-        if hasattr(self._iterable, 'sort'):
-            self._iterable.sort(*args)
-            return self
-        raise TypeError('sort() is only supported for data collection created from list.')
-
-    # def __len__(self):
-    #     """
-    #     Return the length of iterable.
-    #     """
-    #     if not self.is_stream():
-    #         return self._iterable.__len__()
-    #     else:
-    #         raise TypeError('Streamed data collection does not support len.')
-
     @register_dag
     def __add__(self, other):
         """
@@ -944,35 +245,133 @@ class DataCollection(Iterable, DCMixins):
             return repr(self._iterable)
         return super().__repr__()
 
-    @register_dag
-    def head(self, n: int = 5):
+    # Generation Related Function
+    def _factory(self, iterable, parent_stream=True):
         """
-        Get the first n lines of a DataCollection.
+        Factory method for data collection.
+
+        This factory method has been wrapped into a `param_scope()` which contains parent information.
 
         Args:
-            n (`int`):
-                The number of lines to print. Default value is 5.
+            iterable: An iterable object, the data being stored in the DC
+            parent_stream: Whether to copy the parents format (streamed vs unstreamed)
+
+        Returns:
+            DataCollection: DataCollection encapsulating the iterable.
+        """
+        if parent_stream is True:
+            if self.is_stream:
+                if not isinstance(iterable, Iterator):
+                    iterable = iter(iterable)
+            else:
+                if isinstance(iterable, Iterator):
+                    iterable = list(iterable)
+
+        with param_scope() as hp:
+            hp().data_collection.parent = self
+            return DataCollection(iterable)
+
+    @staticmethod
+    @register_dag
+    def range(*arg, **kws):
+        """
+        Generate data collection with ranged numbers.
 
         Examples:
 
-        >>> DataCollection.range(10).head(3).to_list()
-        [0, 1, 2]
+        >>> DataCollection.range(5).to_list()
+        [0, 1, 2, 3, 4]
         """
-
-        def inner():
-            for i, x in enumerate(self._iterable):
-                if i >= n:
-                    break
-                yield x
-
-        return self._factory(inner())
-
-    def run(self):
-        for _ in self._iterable:
-            pass
+        return DataCollection(range(*arg, **kws))
 
     def to_list(self):
         return self._iterable if isinstance(self._iterable, list) else list(self._iterable)
+
+    # Execution Related Function
+    @register_dag
+    def map(self, *arg):
+        """
+        Apply operator to data collection.
+
+        Args:
+            *arg (Callable): functions/operators to apply to data collection;
+
+        Returns:
+            DataCollection: data collections that contains computation results;
+
+        Examples:
+
+        >>> dc = DataCollection([1,2,3,4])
+        >>> dc.map(lambda x: x+1).map(lambda x: x*2).to_list()
+        [4, 6, 8, 10]
+        """
+
+        # mmap
+        if len(arg) > 1:
+            return self.mmap(list(arg))
+        unary_op = arg[0]
+
+        # smap map for stateful operator
+        if hasattr(unary_op, 'is_stateful') and unary_op.is_stateful:
+            return self.smap(unary_op)
+
+        # pmap
+        if self.get_executor() is not None:
+            return self.pmap(unary_op)
+
+        if hasattr(self._iterable, 'map'):
+            return self._factory(self._iterable.map(unary_op))
+
+        if hasattr(self._iterable, 'apply') and hasattr(unary_op, '__dataframe_apply__'):
+            return self._factory(unary_op.__dataframe_apply__(self._iterable))
+
+        # map
+        def inner(x):
+            if isinstance(x, Option):
+                return x.map(unary_op)
+            else:
+                return unary_op(x)
+
+        result = map(inner, self._iterable)
+        return self._factory(result)
+
+    @register_dag
+    def filter(self, unary_op: Callable, drop_empty=False) -> 'DataCollection':
+        """
+        Filter data collection with `unary_op`.
+
+        Args:
+            unary_op (`Callable`):
+                Callable to decide whether to filter the element;
+            drop_empty (`bool`):
+                Drop empty values. Defaults to False.
+
+        Returns:
+            DataCollection: filtered data collection
+        """
+
+        # return filter(unary_op, self)
+        def inner(x):
+            if isinstance(x, Option):
+                if isinstance(x, Some):
+                    return unary_op(x.get())
+                return not drop_empty
+            return unary_op(x)
+
+        if hasattr(self._iterable, 'filter'):
+            return self._factory(self._iterable.filter(unary_op))
+
+        if hasattr(self._iterable, 'apply') and hasattr(unary_op, '__dataframe_filter__'):
+            return DataCollection(unary_op.__dataframe_apply__(self._iterable))
+
+        return self._factory(filter(inner, self._iterable))
+
+    def run(self):
+        """
+        Consume iterables in stream mode.
+        """
+        for _ in self._iterable:
+            pass
 
     def to_df(self):
         """
