+++ towhee/functional/mixins/__init__.py
@@ -27,12 +27,17 @@ from .dag import DagMixin
 from .config import ConfigMixin
 from .compile import CompileMixin
 from .remote import RemoteMixin
+from .list import ListMixin
+from .data_processing import DataProcessingMixin
+from .stream import StreamMixin
+from .safe import SafeMixin
 
 
 class DCMixins(DatasetMixin, DispatcherMixin, DisplayMixin, ParallelMixin,
                ComputerVisionMixin, StateMixin, MetricMixin, RayMixin,
                ServeMixin, MilvusMixin, DagMixin, FaissMixin, ConfigMixin,
-               CompileMixin, RemoteMixin):
+               CompileMixin, RemoteMixin, ListMixin, DataProcessingMixin,
+               SafeMixin, StreamMixin):
 
     def __init__(self) -> None:  # pylint: disable=useless-super-delegation
         super().__init__()
